# openapi.yaml
openapi: 3.1.0
info:
  title: arca API
  version: v1
servers:
- url: https://arca-backend.dcism.org
  description: Production
- url: http://localhost:20255
  description: Local development
paths:
...
  /api/v1/user/{userId}:
    get:
      tags:
      - user-controller
      operationId: getUserDetails
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
      - user-controller
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        "200":
          description: User updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /api/v1/subject/{subjectId}:
    get:
      tags:
      - subject-controller
      operationId: getSubjectDetails
      parameters:
      - name: subjectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectResponse"
    put:
      tags:
      - subject-controller
      operationId: updateSubject
      parameters:
      - name: subjectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectResponse"
    delete:
      tags:
      - subject-controller
      operationId: deleteSubject
      parameters:
      - name: subjectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/schools/edit/{id}:
    put:
      tags:
      - school-controller
      operationId: editSchool
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SchoolResponse"
  /api/v1/posts/{rowId}:
    get:
      tags:
      - post-controller
      operationId: getPost
      parameters:
      - name: rowId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PostResponse"
    put:
      tags:
      - post-controller
      operationId: updatePost
      parameters:
      - name: rowId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PostUpdateResponse"
    delete:
      tags:
      - post-controller
      operationId: deletePost
      parameters:
      - name: rowId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "501":
          description: Post deletion is not implemented
...
  /api/v1/schools:
    get:
      tags:
      - school-controller
      operationId: getAllSchools
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchoolResponse"
    post:
      tags:
      - school-controller
      operationId: addSchool
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SchoolResponse"
  /api/v1/posts:
    get:
      tags:
      - post-controller
      operationId: getAllPosts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
    post:
      tags:
      - post-controller
      operationId: createPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostRequest"
        required: true
      responses:
        "201":
          description: Post created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PostCreateResponse"
  /api/v1/posts/{postId}/approve:
    post:
      tags:
      - post-controller
      operationId: approvePost
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostApprovalRequest"
        required: true
      responses:
        "200":
          description: Post moderation completed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/files/upload:
    post:
      tags:
      - file-controller
      operationId: uploadFile
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: post_id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MultipleFileUploadRequest"
...
  /api/v1/posts/user/{userId}:
    get:
      tags:
      - post-controller
      operationId: getAllUserPost
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
  /api/v1/posts/subject/{subjectId}:
    get:
      tags:
      - post-controller
      operationId: getPostsBySubject
      parameters:
      - name: subjectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
...
  /api/v1/posts/pending/department/{departmentId}:
    get:
      tags:
      - post-controller
      operationId: getPendingPostsByDepartment
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
  /api/v1/posts/mine:
    get:
      tags:
      - post-controller
      operationId: getMyPosts
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
  /api/v1/posts/history/{postId}:
    get:
      tags:
      - post-controller
      operationId: getPostHistory
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
  /api/v1/posts/department/{departmentId}:
    get:
      tags:
      - post-controller
      operationId: getAllDepartmentPost
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"
  /api/v1/files/{id}:
    get:
      tags:
      - file-controller
      operationId: getFileInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FileUploadRequest"
  /api/v1/files/download/{id}:
    get:
      tags:
      - file-controller
      operationId: downloadFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
...
  /api/v1/admin/users/{id}:
    get:
      tags:
      - admin-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
      - admin-controller
      operationId: deleteUser_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /api/v1/schools/delete/{id}:
    delete:
      tags:
      - school-controller
      operationId: deleteSchool
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    UserUpdateRequest:
      type: object
      properties:
        bio:
          type: string
        course:
          type: string
        department:
          type: string
        first_name:
          type: string
        last_name:
          type: string
...
    SubjectRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        departmentIds:
          type: array
          items:
            type: integer
            format: int64
    DepartmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SubjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentResponse"
    SchoolRequest:
      type: object
      properties:
        name:
          type: string
    SchoolResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentResponse"
    PostUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 30
          minLength: 0
        content:
          type: string
        postTag:
          type: string
    PostUpdateResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        status:
          type: string
        message:
          type: string
        post_id:
          type: integer
          format: int32
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    PostRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 30
          minLength: 0
        content:
          type: string
          minLength: 1
        userId:
          type: integer
          format: int64
        departmentId:
          type: integer
          format: int64
        postTag:
          type: string
      required:
      - content
      - departmentId
      - title
    PostCreateResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        post_id:
          type: integer
          format: int32
        message:
          type: string
    PostApprovalRequest:
      type: object
      properties:
        approved:
          type: boolean
        rejectionReason:
          type: string
      required:
      - approved
    FileUploadRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        message:
          type: string
    MultipleFileUploadRequest:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileUploadRequest"
        message:
          type: string
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        course:
          type: string
        department:
          type: string
        bio:
          type: string
        roles:
          type: array
          items:
            type: string
        first_name:
          type: string
        last_name:
          type: string
        profile_picture:
          type: string
...
    FileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
          format: int64
        downloadUrl:
          type: string
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int32
        version:
          type: integer
          format: int32
        title:
          type: string
        content:
          type: string
        status:
          type: string
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        departmentId:
          type: integer
          format: int64
        departmentName:
          type: string
        postTag:
          type: string
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/PostSubjectResponse"
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        upvoteCount:
          type: integer
          format: int32
        downvoteCount:
          type: integer
          format: int32
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileResponse"
        is_latest_version:
          type: boolean
    PostSubjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...