# src/main/java/com/popman/arca/dto/v1/post/PostSubjectResponse.java
package com.popman.arca.dto.v1.post;

public class PostSubjectResponse {
    private Long id;
    private String name;

    public PostSubjectResponse() {}

    public PostSubjectResponse(Long id, String name) {
        this.id = id;
        this.name = name;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}